BITS package, GRAPH package
I've started work on the datastructure for graphs, and for world
entities. The general idea is to keep the entities small, and the graph
fast - since the graph will not be changed at run time, this shouldn't
be too impossible.

I've also worked on trying to increase encapsulation, specificially
between the UI and the event system.